refactor(GameState): extract common code from MainMenu and HeadlessSetup (#4809)
* refactor(GameState): extract common code from MainMenu and HeadlessSetup
* chore(StateLoading): address lint messages 🚮
