+++ engine/src/main/java/org/terasology/engine/core/subsystem/headless/mode/StateHeadlessSetup.java
@@ -6,42 +6,23 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.engine.config.Config;
 import org.terasology.engine.config.WorldGenerationConfig;
-import org.terasology.engine.context.Context;
-import org.terasology.engine.core.ComponentSystemManager;
 import org.terasology.engine.core.GameEngine;
 import org.terasology.engine.core.LoggingContext;
 import org.terasology.engine.core.SimpleUri;
 import org.terasology.engine.core.TerasologyConstants;
-import org.terasology.engine.core.bootstrap.EntitySystemSetupUtil;
-import org.terasology.engine.core.modes.GameState;
+import org.terasology.engine.core.modes.AbstractState;
 import org.terasology.engine.core.modes.StateLoading;
 import org.terasology.engine.core.module.ModuleManager;
 import org.terasology.engine.core.module.StandardModuleExtension;
-import org.terasology.engine.entitySystem.entity.EntityRef;
-import org.terasology.engine.entitySystem.entity.internal.EngineEntityManager;
-import org.terasology.engine.entitySystem.event.internal.EventSystem;
 import org.terasology.engine.game.GameManifest;
 import org.terasology.engine.input.InputSystem;
-import org.terasology.engine.logic.console.Console;
-import org.terasology.engine.logic.console.ConsoleImpl;
-import org.terasology.engine.logic.console.ConsoleSystem;
-import org.terasology.engine.logic.console.commands.CoreCommands;
-import org.terasology.engine.logic.players.LocalPlayer;
-import org.terasology.engine.network.ClientComponent;
 import org.terasology.engine.network.NetworkMode;
-import org.terasology.engine.recording.DirectionAndOriginPosRecorderList;
-import org.terasology.engine.recording.RecordAndReplayCurrentStatus;
-import org.terasology.engine.registry.CoreRegistry;
-import org.terasology.engine.rendering.nui.NUIManager;
-import org.terasology.engine.rendering.nui.internal.NUIManagerInternal;
-import org.terasology.engine.rendering.nui.internal.TerasologyCanvasRenderer;
 import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameInfo;
 import org.terasology.engine.rendering.nui.layers.mainMenu.savedGames.GameProvider;
 import org.terasology.engine.world.internal.WorldInfo;
 import org.terasology.engine.world.time.WorldTime;
 import org.terasology.gestalt.module.Module;
 import org.terasology.gestalt.naming.Name;
-import org.terasology.nui.canvas.CanvasRenderer;
 
 import java.util.List;
 
@@ -49,14 +30,9 @@ import java.util.List;
  * The class is game selection menu replacement for the headless server.
  *
  */
-public class StateHeadlessSetup implements GameState {
+public class StateHeadlessSetup extends AbstractState {
 
     private static final Logger logger = LoggerFactory.getLogger(StateHeadlessSetup.class);
-
-    private EngineEntityManager entityManager;
-    private EventSystem eventSystem;
-    private ComponentSystemManager componentSystemManager;
-    private Context context;
 
     public StateHeadlessSetup() {
     }
@@ -64,31 +40,10 @@ public class StateHeadlessSetup implements GameState {
     @Override
     public void init(GameEngine gameEngine) {
         context = gameEngine.createChildContext();
-        CoreRegistry.setContext(context);
+        initEntityAndComponentManagers();
+        createLocalPlayer(context);
 
-        // let's get the entity event system running
-        EntitySystemSetupUtil.addEntityManagementRelatedClasses(context);
-        entityManager = context.get(EngineEntityManager.class);
-
-        eventSystem = context.get(EventSystem.class);
-        context.put(Console.class, new ConsoleImpl(context));
-
-        NUIManager nuiManager = new NUIManagerInternal((TerasologyCanvasRenderer) context.get(CanvasRenderer.class), context);
-        context.put(NUIManager.class, nuiManager);
-
-        componentSystemManager = new ComponentSystemManager(context);
-        context.put(ComponentSystemManager.class, componentSystemManager);
-
-        componentSystemManager.register(new ConsoleSystem(), "engine:ConsoleSystem");
-        componentSystemManager.register(new CoreCommands(), "engine:CoreCommands");
         componentSystemManager.register(context.get(InputSystem.class), "engine:InputSystem");
-
-        EntityRef localPlayerEntity = entityManager.create(new ClientComponent());
-        LocalPlayer localPlayer = new LocalPlayer();
-        localPlayer.setRecordAndReplayClasses(context.get(DirectionAndOriginPosRecorderList.class), context.get(RecordAndReplayCurrentStatus.class));
-        context.put(LocalPlayer.class, localPlayer);
-        localPlayer.setClientEntity(localPlayerEntity);
-
         componentSystemManager.initialise();
 
         GameManifest gameManifest;
@@ -148,15 +103,6 @@ public class StateHeadlessSetup implements GameState {
     }
 
     @Override
-    public void dispose(boolean shuttingDown) {
-        eventSystem.process();
-
-        componentSystemManager.shutdown();
-
-        entityManager.clear();
-    }
-
-    @Override
     public void handleInput(float delta) {
     }
 
@@ -177,10 +123,5 @@ public class StateHeadlessSetup implements GameState {
     @Override
     public String getLoggingPhase() {
         return LoggingContext.INIT_PHASE;
-    }
-
-    @Override
-    public Context getContext() {
-        return context;
     }
 }
